requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
pydantic-settings
python-dotenv
python-multipart
passlib[bcrypt]
alembic